+++ lib/twenv/commands/write_tweet.rb
@@ -34,7 +34,7 @@ class TWEnv::WriteTweet < TWEnv::Command
   def delay_tweet(tweet)
     time = Time.now + delay
     line.ok("tweet will be published at around #{bold(time.iso8601)}, " \
-            "as long as this twenv.rb process does not exit").end
+            "as long as this twenv.rb process does not exit.").end
     Thread.new do
       sleep delay
       post_tweet(tweet, false)